pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipeforge"
dynamic = ["version"]
description = "CLI toolkit for transpiling CI pipeline specs (Bamboo, Bitbucket Pipelines, GitLab, GitHub Actions)"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "U.N. Owen", email = "void@some.where" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pyyaml>=6.0",
  "click>=8.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
pipeforge = "pipeforge.cli:app"

[project.urls]
Documentation = "https://github.com/U.N. Owen/pipeforge#readme"
Issues = "https://github.com/U.N. Owen/pipeforge/issues"
Source = "https://github.com/U.N. Owen/pipeforge"

[tool.hatch.version]
path = "src/pipeforge/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/pipeforge tests}"

[tool.coverage.run]
source_pkgs = ["pipeforge", "tests"]
branch = true
parallel = true
omit = [
  "src/pipeforge/__about__.py",
]

[tool.coverage.paths]
pipeforge = ["src/pipeforge", "*/pipeforge/src/pipeforge"]
tests = ["tests", "*/pipeforge/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
